Adjust module hierarchy, update elf.h to libc 2.26 version 	modified:   elf.h 	copied:     elf.h -> elf.h.tmp 	modified:   src/error.rs 	new file:   src/format/bindings.rs 	modified:   src/format/elf.rs 	new file:   src/format/executable.rs 	deleted:    src/format/macho.rs 	modified:   src/format/mod.rs 	deleted:    src/format/pe.rs 	modified:   src/lib.rs 	modified:   src/parser/elf_parser.rs 	deleted:    src/structure/bindings.rs 	deleted:    src/structure/elf.rs 	deleted:    src/structure/mod.rs

+++ src/error.rs
@@ -25,13 +25,13 @@ pub enum RustepErrorKind {
     #[fail(display = "Not enough byte, unknown bytes needed")]
     IncompleteUnknown,
     #[fail(display = "Segment type {} not resolved", _0)]
-    SegmentType(usize),
+    SegmentType(u64),
     #[fail(display = "Section type {} not resolved", _0)]
-    SectionType(usize),
+    SectionType(u64),
     #[fail(display = "Segment flag {} invalid", _0)]
-    SegmentFlag(usize),
+    SegmentFlag(u64),
     #[fail(display = "Section flag {} invalid", _0)]
-    SectionFlag(usize),
+    SectionFlag(u64),
 }
 
 impl Fail for RustepError {
